Move hash classes from include to lib
