+++ pulsar-client-cpp/lib/Hash.h
