+++ pulsar-client-cpp/lib/Murmur3_32Hash.cc
@@ -21,7 +21,7 @@
 // The original MurmurHash3 was written by Austin Appleby, and is placed in the
 // public domain. This source code, implemented by Licht Takeuchi, is based on
 // the orignal MurmurHash3 source code.
-#include <pulsar/Murmur3_32Hash.h>
+#include "Murmur3_32Hash.h"
 
 #include <boost/predef.h>
